requirements.txt
streamlit
requests
streamlit-lottie
pandas
numpy
joblib
deepface
Pillow
librosa
soundfile
matplotlib
tensorflow
plotly
opencv-python
audio-recorder-streamlit
streamlit-extras
altair
langchain
ollama
